Refactor streams out of nu-cli into a separate subcrate
