+++ crates/nu-cli/src/shell/shell_manager.rs
@@ -8,7 +8,6 @@ use crate::commands::rm::RemoveArgs;
 use crate::prelude::*;
 use crate::shell::filesystem_shell::FilesystemShell;
 use crate::shell::shell::Shell;
-use crate::stream::OutputStream;
 use nu_errors::ShellError;
 use parking_lot::Mutex;
 use std::error::Error;
