+++ crates/nu-streams/src/lib.rs
@@ -1,7 +1,9 @@
 mod input;
 mod interruptible;
 mod output;
+mod traits;
 
 pub use input::*;
 pub use interruptible::*;
 pub use output::*;
+pub use traits::*;
